Replaced byte[] with int[]

+++ Accounting/UTXO/Block.cs
@@ -13,9 +13,9 @@ namespace BToken.Accounting
     public UInt256 HeaderHash;
     public NetworkHeader Header;
     public List<TX> TXs;
-    public List<byte[]> TXHashes;
+    public List<int[]> TXHashes;
 
-    public Block(NetworkHeader header, UInt256 headerHash, List<TX> tXs, List<byte[]> tXHashes)
+    public Block(NetworkHeader header, UInt256 headerHash, List<TX> tXs, List<int[]> tXHashes)
     {
       Header = header;
       HeaderHash = headerHash;
